Fix #613: partition-aware bone-index remap on BsTriShape skin import
Pre-fix `extract_skin_bs_tri_shape` cloned the inline `[u8; 4]`
bone_indices verbatim and treated them as global indices into the
skin's bone list. Per Bethesda's BSTriShape format the per-vertex
indices are partition-LOCAL — each `NiSkinPartition.partitions[i]`
carries its own `bones: Vec<u16>` palette and the inline indices
point into that palette. Multi-partition shapes (Skyrim
Argonian/Khajiit body + worn armour merging into the partition list,
modded > 256-bone skins) silently aliased every vertex past
partition 0 to whichever bone happened to sit at the partition-local
slot — limbs followed wrong joints, hair bound to torso bones.

Two coupled changes:

1. **Promote `ImportedSkin.vertex_bone_indices`** from `Vec<[u8; 4]>`
   to `Vec<[u16; 4]>`. Skyrim character + worn-armour skins routinely
   approach 255 bones; mods exceed it. The renderer's `Vertex` struct
   already carries `[u32; 4]` so the existing widening cast at
   `byroredux/src/scene.rs:1093` keeps working — just promotes from a
   wider source type.

2. **Add `remap_bs_tri_shape_bone_indices`** that walks
   `shape.skin_ref` → `NiSkinInstance` → `skin_partition_ref` →
   `NiSkinPartition`, builds an inverse `vertex_map` lookup
   (global vertex → owning partition), and remaps each vertex's
   four u8 partition-local indices via `partition.bones[local_idx]`.
   Single-partition shapes take an identity-widen fast path.
   Missing / null partitions fall back to identity widen with a
   warning when `bone_refs.len() > u16::MAX`.

Bonus: `densify_sparse_weights` (the legacy NiTriShape path) had the
same u8 ceiling — bones past index 255 were silently dropped. Widened
to `u16` with `u16::MAX` as the new vacant-slot sentinel.

3 regression tests:
- `multi_partition_remap_picks_correct_global_per_vertex` — synthetic
  2-vertex / 2-partition shape with distinct palettes; verifies that
  partition-local `[0, 1, 0, 1]` maps to globals `[2, 3, 2, 3]` for
  vertex 0 and `[1, 3, 1, 3]` for vertex 1 (pre-fix would clone
  `[0, 1, 0, 1]` for both).
- `single_partition_shape_widens_indices_directly` — fast-path lock
  so single-partition meshes don't pay the remap walk.
- `missing_skin_partition_falls_back_to_identity_widen` — defensive
  path when the partition table is null (synthetic / mod content).

Real-data check: vanilla `Skyrim - Meshes0.bsa` parse rate stays at
100% (18862/18862 clean), proving the remap doesn't break shapes
that previously parsed cleanly through the silent-alias path.

+++ crates/nif/src/import/mod.rs
@@ -377,9 +377,19 @@ pub struct ImportedSkin {
     /// Skeleton root bone name, if identifiable. The animation system
     /// uses this to know where to start applying joint transforms.
     pub skeleton_root: Option<Arc<str>>,
-    /// Per-vertex bone indices (up to 4). Parallel to `ImportedMesh::positions`.
-    /// Empty if weights come from a separate source (BSTriShape).
-    pub vertex_bone_indices: Vec<[u8; 4]>,
+    /// Per-vertex bone indices (up to 4) — **already remapped to
+    /// global indices into [`ImportedSkin::bones`]**. Parallel to
+    /// `ImportedMesh::positions`. Pre-#613 BsTriShape stored these as
+    /// `[u8; 4]` carrying *partition-local* values (indices into the
+    /// per-`NiSkinPartition` `bones` palette, not the global skin
+    /// list); shapes with > 1 partition silently aliased every vertex
+    /// past partition 0 to the wrong bones. The importer now walks the
+    /// partition table during extraction and emits global indices, so
+    /// every value here resolves directly through `bones[idx]`.
+    /// Widened to `u16` because vanilla Skyrim character + worn-armor
+    /// skins routinely exceed 255 bones; mods can push higher.
+    /// Empty if weights come from a separate source.
+    pub vertex_bone_indices: Vec<[u16; 4]>,
     /// Per-vertex bone weights (up to 4). Must sum to 1.0 after
     /// normalization. Parallel to `vertex_bone_indices`.
     pub vertex_bone_weights: Vec<[f32; 4]>,
